Move Autocorrection Program to Github
As a final project for my Artificial intelligence class, I wrote a program similar to iMessages auto correction and auto suggestion. Model can be train on any set of text you want and actively learns from the user input. That is to say, as you type it is reading you keystrokes and adjusting its prediction based on you personal text pattern. If you go back and delete a word or character it also "unlearns".